requirements.txt
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
apprise==1.9.2
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==25.1.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coolname==2.2.0
croniter==6.0.0
cryptography==44.0.0
dateparser==1.2.0
deprecated==1.2.17
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.7
fsspec==2024.12.0
future==1.0.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.5
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.5.0
-e file:///home/sohan/IndiaML/sohan/IndiaML/indiaml
jinja2==3.1.5
jinja2-humanize-extension==0.4.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mako==1.3.8
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
openai==1.60.1
openreview-py==1.46.2
opentelemetry-api==1.29.0
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
pendulum==3.0.0
pip==24.3.1
prefect==3.1.14
prometheus-client==0.21.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pygments==2.19.1
pyjwt==2.10.1
pylatexenc==2.10
pymupdf==1.25.2
pymupdf4llm==0.0.17
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.6.1
pytz==2024.2
pyyaml==6.0.2
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.22.3
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.37
starlette==0.45.3
text-unidecode==1.3
time-machine==2.16.0
tld==0.13
toml==0.10.2
tqdm==4.67.1
typer==0.13.1
typing-extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==13.1
wrapt==1.17.2
zipp==3.21.0
